Starting to work on preparing user interface.

+++ service/src/main/java/pl/cinemaproject/service/AdminService.java
@@ -1,5 +1,6 @@
 package pl.cinemaproject.service;
 
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import pl.cinemaproject.persistence.model.Cinema;
 import pl.cinemaproject.persistence.model.CinemaRoom;
@@ -8,6 +9,7 @@ import pl.cinemaproject.repository.CinemaRepository;
 import pl.cinemaproject.repository.CinemaRoomRepository;
 import pl.cinemaproject.repository.CityRepository;
 import pl.cinemaproject.repository.SeatRepository;
+import pl.cinemaproject.service.exception.AdminServiceException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -17,20 +19,17 @@ import java.util.stream.Collectors;
 public class AdminService {
 
 
-
     private final CityRepository cityRepository;
     private final CinemaRepository cinemaRepository;
     private final CinemaRoomRepository cinemaRoomRepository;
     private final SeatRepository seatRepository;
 
 
+    public String addNewCity(@NonNull String cityName) {
 
 
-
-    public String addNewCity(String cityName){
-
-        if(cityRepository.findByName(cityName).isPresent()){
-            throw new NullPointerException("City " + cityName + " already exists");
+        if (cityRepository.findByName(cityName).isPresent()) {
+            throw new AdminServiceException(cityName + " already exists in database");
         }
 
         return cityRepository
@@ -44,10 +43,10 @@ public class AdminService {
     }
 
 
-    public int findCityIdByName(String cityName) {
+    public int findCityIdByName(@NonNull String cityName) {
 
-        if(cityRepository.findByName(cityName).isEmpty()){
-            throw new NullPointerException("City " + cityName + " dose not exist");
+        if (cityRepository.findByName(cityName).isEmpty()) {
+            throw new AdminServiceException(cityName + " dose not exist in data base");
         }
 
         return cityRepository.findByName(cityName).map(City::getId).orElseThrow();
@@ -55,62 +54,71 @@ public class AdminService {
     }
 
 
-    public String addNewCinema(Cinema cinema){
+    public String addNewCinema(@NonNull Cinema cinema) {
 
-        if(cinemaRepository.findByName(cinema.getName()).isPresent()){
-            throw new NullPointerException("Cinema " + cinema.getName() + " already exists");
+        if (cinemaRepository.findByName(cinema.getName()).isPresent()) {
+            throw new AdminServiceException(cinema.getName() + " already exists in data base");
         }
 
         return cinemaRepository.add(cinema).orElseThrow().getName();
     }
 
-    public int findCinemaIdByName(String cinemaName){
+    public int findCinemaIdByName(@NonNull String cinemaName) {
 
-        if(cinemaRepository.findByName(cinemaName).isEmpty()){
-            throw new NullPointerException("City " + cinemaName + " dose not exist");
+        if (cinemaRepository.findByName(cinemaName).isEmpty()) {
+            throw new AdminServiceException(cinemaName + " dose not exist in data base");
         }
 
         return cinemaRepository.findByName(cinemaName).map(Cinema::getId).orElseThrow();
     }
 
-    public int findCinemaRoomIdByName(String cinemaRoomName, String cinemaName){
+    public int findCinemaRoomIdByName(@NonNull String cinemaRoomName, @NonNull String cinemaName) {
 
         var cinemaId = findCinemaIdByName(cinemaName);
 
-        return cinemaRoomRepository.findCinemaRoomByNameWithCinemaId(cinemaRoomName,cinemaId).orElseThrow().getId();
+        return cinemaRoomRepository.findCinemaRoomByNameWithCinemaId(cinemaRoomName, cinemaId).orElseThrow().getId();
 
     }
 
 
-    public String addNewCinemaRoom(CinemaRoom cinemaRoom){
+    public String addNewCinemaRoom(@NonNull CinemaRoom cinemaRoom) {
 
-        if (cinemaRoomRepository.findByName(cinemaRoom.getName()).isPresent()){
-            throw new NullPointerException("City " + cinemaRoom + " dose not exist");
+        if (cinemaRoomRepository.findCinemaRoomByNameWithCinemaId(cinemaRoom.getName(),
+                cinemaRoom.getCinemaId()).isPresent()) {
+            throw new AdminServiceException(cinemaRoom.getName() + " dose already exist");
         }
 
         return cinemaRoomRepository.add(cinemaRoom).orElseThrow().getName();
     }
 
 
-    public String updateCity(City city){
+    public String updateCity(@NonNull City city) {
 
-//        var cityToUpdate = cityRepository.findByName(cityName).orElseThrow();
-//        cityToUpdate.setName("New");
-        return cityRepository.update(city,city.getId()).orElseThrow().getName();
+        if(cityRepository.findByName(city.getName()).isEmpty()){
+         throw new AdminServiceException(city.getName() + "dose not exist in data base");
+
+        }
+        return cityRepository.update(city, city.getId()).orElseThrow().getName();
 
 
     }
 
-    public String updateCinema(Cinema cinema){
+    public String updateCinema(@NonNull Cinema cinema) {
 
+        if(cinemaRepository.findByName(cinema.getName()).isEmpty()){
+            throw new AdminServiceException(cinema.getName() + "dose not exist in data base");
 
+        }
         return "New cinema name " + cinemaRepository.update(cinema, cinema.getId()).orElseThrow().getName();
 
 
     }
 
-    public String removeCity(String cityName){
+    public String removeCity(@NonNull String cityName) {
 
+        if (cityRepository.findByName(cityName).isEmpty()){
+            throw new AdminServiceException(cityName + "dose not exist in data base");
+        }
         var cityToRemove = findCityIdByName(cityName);
         return "Successfully remove " + cityRepository.deleteById(cityToRemove).orElseThrow().getName();
     }
@@ -118,6 +126,9 @@ public class AdminService {
 
     public String removeCinema(String cinemaName) {
 
+        if (cinemaRepository.findByName(cinemaName).isEmpty()){
+            throw new AdminServiceException(cinemaName + "dose not exist in data base");
+        }
         var cinemaToRemove = findCinemaIdByName(cinemaName);
         return "Successfully remove " + cinemaRepository.deleteById(cinemaToRemove).orElseThrow().getName();
     }
@@ -125,16 +136,19 @@ public class AdminService {
 
     public String removeCinemaRoom(String cinemaRoomName, String cinemaName) {
 
-        var cinemaRoomToRemove = findCinemaRoomIdByName(cinemaRoomName,cinemaName);
+        if (findCinemaRoomIdByName(cinemaRoomName,cinemaName) == 0){
+            throw new AdminServiceException(cinemaRoomName + "dose not exist in data base");
+        }
+
+        var cinemaRoomToRemove = findCinemaRoomIdByName(cinemaRoomName, cinemaName);
 
         return "Successfully remove " + cinemaRoomRepository.deleteById(cinemaRoomToRemove).orElseThrow().getName();
-     }
-
+    }
 
 
     //ToDo Zapytaj Krzyska czy tu powinnienem użyć strumini czy żeby zostawić zabawę
     // bazie danych powinno już w sql pobierać tylko tabele z nazwami miast
-    public List<String> getAllCitiesNames(){
+    public List<String> getAllCitiesNames() {
 
         return
                 cityRepository
@@ -145,12 +159,12 @@ public class AdminService {
 
     }
 
-    public Cinema getCinemaByName(String cinemaName){
+    public Cinema getCinemaByName(String cinemaName) {
 
         return cinemaRepository.findByName(cinemaName).orElseThrow();
     }
 
-    public List<String> getAllCinemasNames(){
+    public List<String> getAllCinemasNames() {
 
         return cinemaRepository
                 .findAll()
@@ -160,7 +174,7 @@ public class AdminService {
     }
 
 
-    public List<String> getAllCinemaRoomsForOneCinema(int id){
+    public List<String> getAllCinemaRoomsForOneCinema(int id) {
 
         return cinemaRoomRepository
                 .findAllCinemaRoomsByCinemaId(id)
@@ -170,16 +184,19 @@ public class AdminService {
 
     }
 
-    public CinemaRoom getCinemaRoomByName(String cinemaRoomName){
+    public CinemaRoom getCinemaRoomByName(@NonNull String cinemaRoomName) {
 
         return cinemaRoomRepository
                 .findByName(cinemaRoomName)
                 .orElseThrow();
     }
 
-    public String updateCinemaRoom(CinemaRoom cinemaRoom) {
+    public String updateCinemaRoom(@NonNull CinemaRoom cinemaRoom) {
 
-
+        if (cinemaRoomRepository.findCinemaRoomByNameWithCinemaId(cinemaRoom.getName(),
+                cinemaRoom.getCinemaId()).isEmpty()){
+            throw new AdminServiceException(cinemaRoom.getName() + "dose not exist in data base");
+        }
         return "New cinema room name is " + cinemaRoomRepository.update(cinemaRoom, cinemaRoom.getId()).orElseThrow().getName();
     }
